Party Invitations
invitations can be sent out for existing users (public and private parties) or to non-users by entering a list of email addresses (private parties only)

Invitation emails will include a link for the recipient to click.  This link will redirect them to a page which will confirm that they have been RSVPd to the related party.