changed confirmation and fixed user create to accept address
Confirmations are now simpler:  A user, after signing up, will receive an email with a temporary password.  It is the client's job to insist that the user change their password upon logging in for the first time before doing anything else in the app.

Also, registrations controller was removed and the user creation functionality is now in the users controller.  We can now post the address along with the user for creation and it will be done in one go instead of having to create the address separately.