Facebook API using koala
sessions controller tries to log the user in based on the access token received.  If there is no user in the system that matches the UID that we get back from FB, an invalid credentials is returned.

Registrations controller will take a Facebook access token and create a Facebook user instead of one based on an email and password.