Cleaned up serializers
Also changed teams and sports to followed_teams and followed_sports on the user and venue.

Moved the following of teams and sports out of the teams and sports controllers in to the users controller, which makes permissions much easier.

Added timestamps to a couple migrations which were missing it.